requirements.txt
customtkinter>=5.2.0
yt-dlp>=2024.1.0
spotdl>=4.2.0
mutagen>=1.47.0
Pillow>=10.0.0
pygame>=2.5.0
requests>=2.31.0
packaging>=23.0
gitpython>=3.1.0
pyinstaller>=6.0.0
darkdetect
CTkMessagebox
pillow-avif-plugin
